Fix flawed assumption in includes/excludes predicate
Fixes an issue caused by the assumption that ordinal/categorical node
values are always arrays. In fact they can be strings when a single
value is assigned on categorical/ordinal bin interface.

This commit fixes the issue by checking if the value is an array before
calling array specific methods such as `.some()`.

Adds test coverage for the scenario where values are strings or numbers.