Fix includes/excludes for string and integer variable values
App crashed when variable value was an integer, and app was incorrectly
determining includes/excludes when variable value was a string.

This happened because previous code assumed that variable values were
always arrays, and called 'Array.includes()' on them. This worked for
string values (albeit returning incorrect results), but crashed for
integer values.