Allow multiple values for exactly/not predicate
This PR adds the ability to provide multiple values to the `exactly` and
`not` predicates. This applies to categorical variables only.

The comparison is done using the `isEqual` lodash method, with sorting
applied to the values before comparison. This means that the order of
the values provided to the predicate does not matter.

Architect will need to be updated to produce rules that use this new
functionality.